Adding trendlines to scatter
Think this should do it - not sure what the issue was.  Only thing this doesn't include is the scale_colour_hue thing - not sure what that was trying to accomplish.

I also added a second set of paths at the beginning, one set for my computer and one for yours so that we dont have to keep changing them, just comment and uncomment.